fix(product): keep the cloud poll plan out of the login first-load graph (PRO-230)
`useAwaitingCloudWorkspaceEntries` has exactly one consumer — the poll loop —
but it sat in `use-pending-workspace-entries.ts`, which the signed-out shell
reaches eagerly through the workspace navigation workflow. That dragged
`cloud-workspace-poll-plan` into the login entry chunk for nothing.

Move the selector into `use-cloud-workspace-polling.ts` beside its only caller.
The shared readers keep only what the shell actually reads.
