+++ apps/packages/product-client/src/hooks/workspaces/derived/use-pending-workspace-entries.ts
@@ -8,12 +8,7 @@ import {
   pendingWorkspaceEntries,
   pendingWorkspaceEntry,
 } from "#product/lib/domain/workspaces/creation/pending-entry-registry";
-import {
-  isAwaitingCloudWorkspaceEntry,
-} from "#product/lib/domain/workspaces/cloud/cloud-workspace-poll-plan";
 import { useSessionSelectionStore } from "#product/stores/sessions/session-selection-store";
-
-const EMPTY_PENDING_WORKSPACE_ENTRIES: readonly PendingWorkspaceEntry[] = [];
 
 /** The pending attempt the user is looking at, if any. Never stored. */
 export function useAttendedPendingWorkspaceEntry(): PendingWorkspaceEntry | null {
@@ -33,19 +28,6 @@ export function useAttendedPendingWorkspaceEntry(): PendingWorkspaceEntry | null
       selection.selectedWorkspaceId,
     ],
   );
-}
-
-/**
- * Every attempt parked on cloud provisioning, attended or not: the polling
- * loop drives all of them so a launch completes while the user is elsewhere
- * (PRO-230).
- */
-export function useAwaitingCloudWorkspaceEntries(): readonly PendingWorkspaceEntry[] {
-  const registry = useSessionSelectionStore((state) => state.pendingWorkspaces);
-  return useMemo(() => {
-    const entries = pendingWorkspaceEntries(registry).filter(isAwaitingCloudWorkspaceEntry);
-    return entries.length > 0 ? entries : EMPTY_PENDING_WORKSPACE_ENTRIES;
-  }, [registry]);
 }
 
 export function usePendingWorkspaceEntries(): readonly PendingWorkspaceEntry[] {
